Move the BoxCloneSyncService to mas-http
